Revert "v0.2.0 is ready (#65)"
This reverts commit 2172f347cc23df417d13e04aadacdf182cdc678d.
