+++ .github/workflows/docker_workflow.yml
@@ -1,17 +1,17 @@
-name: Docker CI [Disable CG]
+name: Docker Compose Actions Workflow
 on: push
 jobs:
   docker:
     timeout-minutes: 10
-    runs-on: ubuntu-22.04
-    env:
-      NUOJ_SANDBOX_ENABLE_CG: 0
+    runs-on: ubuntu-20.04
     steps:
     - name: Checkout
       uses: actions/checkout@v2
-    
-    - name: Check config
-      run: docker compose config
+
+    - name: Check Control Group
+      run: |
+        cd /sys/fs/cgroup/memory
+        ls
 
     - name: Start containers
       run: docker compose up --build --exit-code-from "sandbox-test" sandbox-test
