--- .github/workflows/docker_workflow_enable_cg.yml
@@ -1,34 +1,0 @@
-name: Docker CI [Enable CG]
-on: push
-jobs:
-  docker:
-    timeout-minutes: 10
-    runs-on: ubuntu-20.04
-    env:
-      NUOJ_SANDBOX_ENABLE_CG: 1
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v2
-
-    - name: Check Control Group
-      run: |
-        cd /sys/fs/cgroup/memory
-        ls
-    
-    - name: Disable SWAP ASLR THS
-      run: |
-        sudo swapoff -a
-        echo "0" | sudo tee /proc/sys/kernel/randomize_va_space
-        sudo echo "never" | sudo tee /sys/kernel/mm/transparent_hugepage/enabled
-        sudo echo "never" | sudo tee /sys/kernel/mm/transparent_hugepage/defrag
-        sudo echo "0" | sudo tee /sys/kernel/mm/transparent_hugepage/khugepaged/defrag
-    
-    - name: Check config
-      run: docker compose config
-
-    - name: Start containers
-      run: docker compose up --build --exit-code-from "sandbox-test" sandbox-test
-
-    - name: Stop containers
-      if: always()
-      run: docker compose down -v
