+++ backend/env_setting.py
@@ -1,16 +1,7 @@
-#!/bin/python3
 import os
 import subprocess
 import sys
-import apt
 
-
-def check_and_install_env():
-    cache = apt.Cache()
-    if not cache['sysfsutils'].is_installed:
-        print("Please install sysfsutils package first!")
-        print("apt-get install -y sysfsutils")
-        exit(1)
 
 def set_grub():
     f = open("/etc/default/grub", mode="r")
@@ -37,32 +28,42 @@ def set_grub():
     f.close()
 
 
-def set_boot_task():
-    if not subprocess.check_output( ["cat", "/proc/sys/kernel/randomize_va_space"] ).decode("utf-8") == "0\n" :
-        os.system("echo 'kernel.randomize_va_space = 0' >> /etc/sysfs.conf")
-    if not subprocess.check_output(["cat", "/sys/kernel/mm/transparent_hugepage/enabled"]).decode("utf-8")== "always madvise [never]\n":
-        os.system("echo 'kernel/mm/transparent_hugepage/enabled = never' >> /etc/sysfs.conf")
-    if not subprocess.check_output(["cat", "/sys/kernel/mm/transparent_hugepage/defrag"]).decode("utf-8")== "always defer defer+madvise madvise [never]\n":
-        os.system("echo 'kernel/mm/transparent_hugepage/defrag = never' >> /etc/sysfs.conf")
-    if not subprocess.check_output(["cat", "/sys/kernel/mm/transparent_hugepage/khugepaged/defrag"]).decode("utf-8")== "0\n":
-        os.system("echo 'kernel/mm/transparent_hugepage/khugepaged/defrag = 0' >> /etc/sysfs.conf")
-    if os.path.exists("/etc/rc.local"):
-        f = open("/etc/rc.local", mode="r")
-        temp = f.readlines()
-        f.close()
-        if not "swapoff -a" in temp:
-            temp.insert(1, "swapoff -a")
-            f = open("/etc/rc.local", mode="w")
-            f.writelines(["#!/bin/bash\n", "swapoff -a\n", "exit 0"])
-            f.close()
-    else:
-        f = open("/etc/rc.local", mode="w")
-        f.writelines(["#!/bin/bash\n", "swapoff -a\n", "exit 0"])
-        f.close()
-    
+def set_kernel():
+    if (
+        not subprocess.check_output(
+            ["cat", "/proc/sys/kernel/randomize_va_space"]
+        ).decode("utf-8")
+        == "0\n"
+    ):
+        os.system("echo 'kernel.randomize_va_space = 0' >> /etc/sysctl.conf")
+    if (
+        not subprocess.check_output(
+            ["cat", "/sys/kernel/mm/transparent_hugepage/enabled"]
+        ).decode("utf-8")
+        == "always madvise [never]\n"
+    ):
+        os.system(
+            "echo 'kernel/mm/transparent_hugepage/enabled = never' >> /etc/sysfs.conf"
+        )
+    if (
+        not subprocess.check_output(
+            ["cat", "/sys/kernel/mm/transparent_hugepage/defrag"]
+        ).decode("utf-8")
+        == "always defer defer+madvise madvise [never]\n"
+    ):
+        os.system(
+            "echo 'kernel/mm/transparent_hugepage/defrag = never' >> /etc/sysfs.conf"
+        )
+    if (
+        not subprocess.check_output(
+            ["cat", "/sys/kernel/mm/transparent_hugepage/khugepaged/defrag"]
+        ).decode("utf-8")
+        == "0\n"
+    ):
+        os.system(
+            "echo 'kernel/mm/transparent_hugepage/khugepaged/defrag = 0' >> /etc/sysfs.conf"
+        )
 
-def set_rc_local_permission():
-    os.chmod("/etc/rc.local", 0o775)
 
 def run_save(reboot=True):
     os.system("sysctl -p")
@@ -80,8 +81,6 @@ if __name__ == "__main__":
             reboot_flag = True
         else:
             reboot_flag = False
-    check_and_install_env()
     set_grub()
-    set_boot_task()
-    set_rc_local_permission()
+    set_kernel()
     run_save(reboot=reboot_flag)
