--- backend/tests/cg/setting/test_setting.py
@@ -1,12 +1,0 @@
-from os import environ
-
-import pytest
-
-from tests.nocg.setting.test_setting import (
-    setting,
-    TestSetting
-)
-
-@pytest.mark.skipif(environ.get("NUOJ_SANDBOX_ENABLE_CG", "0") == "0", reason="No or not enable CG Environment Variable [NUOJ_SANDBOX_ENABLE_CG]")
-class TestSetting(TestSetting):
-    pass
